breakout nested list, fix types
