+++ napari/utils/events/__init__.py
@@ -1,2 +1,3 @@
-from .event import Event, EventEmitter, EmitterGroup
-from .containers._list import EventedList, NestableEventedList
+from .event import EmitterGroup, Event, EventEmitter  # isort:skip
+from .containers._list import EventedList
+from .containers._nested_list import NestableEventedList
